refactor, finish /my-stories
